requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
cryptography==45.0.3
cssselect2==0.8.0
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
Django==5.2.1
djangorestframework==3.16.0
docopt==0.6.2
executing==2.2.0
fastjsonschema==2.21.1
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
openai==1.83.0
oscrypto==1.3.0
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.2.1
pipreqs==0.5.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pyHanko==0.29.0
pyhanko-certvalidator==0.27.0
pypdf==5.5.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
reportlab==4.4.1
requests==2.32.3
rpds-py==0.25.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
stack-data==0.6.3
svglib==1.5.1
tinycss2==1.4.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
-e git+https://github.com/YuliiaSerohina/Dteam-django-practical-test.git@a9fb5013608348973a6a21e4d4a2a720716b8904#egg=vedteam
wcwidth==0.2.13
webencodings==0.5.1
xhtml2pdf==0.2.17
yarg==0.1.9
